follow api - unfollow added, send request to follow noti !

+++ src/main/kotlin/com/example/frontServer/dto/board/response/CommentResult.kt
@@ -26,7 +26,7 @@ data class CommentResult(
         fun of(
             boardWithCommentCount: BoardWithCommentCount,
             writerName: String,
-            likeCount: Int
+            likeCount: Long
         ): CommentResult {
             return CommentResult(
                 id = boardWithCommentCount.board.id!!,
